Add files via upload
The main objective of the system is to establish a complete RFID and Fingerprint Attendance System that will check/monitor the attendance/status (in or out) of each employee. The Raspberry pi acts as a system that will be used as a client for the Raspberry Pi, Fingerprint and RFID Reader. It can also be a data storage and logic master that checks the validity of user’s attendance. Different method of evaluation was done to make sure that the systems functionality.
and design would be consider